requirements.txt
openai==1.12.0
pyttsx3==2.90
SpeechRecognition==3.10.1
pyaudio==0.2.14
numpy==1.26.4
sentence-transformers==2.5.1
python-dotenv==1.0.1
requests==2.31.0
beautifulsoup4==4.12.3
schedule==1.2.1
sounddevice==0.4.6
scipy==1.12.0 